# docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: Controller for card control for users
    description: Operations for common users
  - name: Сontroller for card control by admin
    description: Only admin can do crud operations
paths:
  /admin/{id}:
    get:
      tags:
        - Сontroller for card control by admin
      summary: Get any card by id
      operationId: getCardById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
    put:
      tags:
        - Сontroller for card control by admin
      summary: Update existed card for user
      operationId: updateCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
    delete:
      tags:
        - Сontroller for card control by admin
      summary: Block card by id
      operationId: deleteCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
  /admin/user/{id}:
    get:
      tags:
        - Controller for user control by admin
      summary: Get user by id
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      tags:
        - Controller for user control by admin
      summary: Update new user by id
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      tags:
        - Controller for user control by admin
      summary: Delete user by id
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /cards/transfer:
    post:
      tags:
        - Controller for card control for users
      summary: Transfer between user's own cards
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Transfer"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /auth/login:
    post:
      tags:
        - Login controller
      summary: Creating token for signed up client
      operationId: login
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
        - name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/user:
    get:
      tags:
        - Controller for user control by admin
      summary: Get all existed users and admins
      operationId: getAllUsers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
    post:
      tags:
        - Controller for user control by admin
      summary: Create new user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /admin/newcard:
    post:
      tags:
        - Сontroller for card control by admin
      summary: Create new card for user
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
  /ping:
    get:
      tags:
        - Endpoint for checking service availability
      operationId: ping
      responses:
        "200":
          description: OK
  /cards/{id}:
    get:
      tags:
        - Controller for card control for users
      summary: Get particular user's card by id
      operationId: getMyCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
  /cards/my:
    get:
      tags:
        - Controller for card control for users
      summary: Get all the user's cards
      operationId: getMyCards
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
  /admin:
    get:
      tags:
        - Сontroller for card control by admin
      summary: 'Get all the cards of a specific user '
      operationId: getAllByOwner
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
  /admin/all:
    get:
      tags:
        - Сontroller for card control by admin
      summary: Get all existed cards
      operationId: getAllCards
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
components:
  schemas:
    CardRequest:
      type: object
      properties:
        balance:
          type: number
        ownerName:
          type: string
        number:
          type: string
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardMask:
          type: string
        owner:
          type: string
        balance:
          type: number
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
    UserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
        enable:
          type: boolean
    Transfer:
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number