# src/main/java/com/example/bankcards/controller/AdminUserController.java
package com.example.bankcards.controller;

import com.example.bankcards.dto.user.UserRequest;
import com.example.bankcards.dto.user.UserResponse;
import com.example.bankcards.entity.User;
import com.example.bankcards.service.CardService;
import com.example.bankcards.service.UserService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@Tag(name = "Controller for user control by admin")
@RequestMapping("/admin/user")
public class AdminUserController {

//    private final CardService cardService;
    private final UserService userService;
    private static final Logger log = LoggerFactory.getLogger(AdminUserController.class);

    public AdminUserController(CardService cardService, UserService userService) {
        this.userService = userService;
//        this.cardService = cardService;
    }

    @Operation(summary = "Get all existed users and admins")
    @GetMapping
    public ResponseEntity<List<UserResponse>> getAllUsers() {
        log.info("Called getAllUsers");
        try {
            return ResponseEntity.status(HttpStatus.OK)
                    .body(userService.getAllUsers());
        } catch (Exception e) {
            return ResponseEntity.status(404).build();
        }
    }


    @Operation(summary = "Create new user")
    @PostMapping()
    public ResponseEntity<UserResponse> createUser(@RequestBody UserRequest userRequest) {
        log.info("Called create user: username = {}", userRequest.username());
        try {
            return ResponseEntity.status(HttpStatus.CREATED)
                    .body(userService.createUser(userRequest));
        } catch (Exception e ) {
            return ResponseEntity.status(405).build();
        }
    }

    @Operation(summary = "Update new user by id")
    @PutMapping("/{id}")
    public ResponseEntity<UserResponse> updateUser(@PathVariable Long id, @RequestBody UserRequest userRequest) {
        log.info("Called update user: id = {}", id);
        try {
            return ResponseEntity.status(HttpStatus.OK)
                    .body(userService.updateUser(id, userRequest));
        } catch (Exception e ) {
            return ResponseEntity.status(405).build();
        }
    }

    @Operation(summary = "Delete user by id")
    @DeleteMapping("/{id}")
    public ResponseEntity<UserResponse> deleteUser(@PathVariable Long id) {
        log.info("Called delete user: id = {}", id);
        try {
            return ResponseEntity.status(HttpStatus.OK)
                    .body(userService.deleteUser(id));
        } catch (Exception e) {
            return ResponseEntity.status(404).build();
        }
    }

    @Operation(summary = "Get user by id")
    @GetMapping("/{id}")
    public ResponseEntity<UserResponse> getUserById(@PathVariable Long id) {
        log.info("Called get user: id = {}", id);
        try {
            return ResponseEntity.status(HttpStatus.OK)
                    .body(userService.getUserById(id));
        } catch (Exception e) {
            return ResponseEntity.status(404).build();
        }
    }
}